+++ tests/test_integrations.py
@@ -1,0 +1,9 @@
+"""Test integrations that work with unittest testcases."""
+# pylint: disable=missing-docstring,no-self-use
+import unittest
+
+from tests.features import FlexmockTestCase
+
+
+class TestIntegrations(FlexmockTestCase, unittest.TestCase):  # pylint: disable=too-many-ancestors
+    pass